Removed outdated (or never accepted) #gnome tags
I compared the #gnome tags with the list of official Gnome applications. The result is that I removed many #gnome tags, because the apps or tags are outdated or the applications were never official Gnome applications.

For some I wondered myself, e.g. Gnote or Shotwell, but they are not listed on the official Gnome website.

source:
https://apps.gnome.org/