Added Rollit (please see notes in the description)
- [Rollit](https://github.com/ellie-commons/rollit) - Roll the dice application from six-sided dice up to 999 sides, designed for elementary OS `#vala` `#gtk4` `#granite`.

Note: Rollit and Chance are different pieces of software, but related, as both are based on an old version of chance. Chance has been rewritten in rust for gnome systems, while rollit has been continued for elementary OS using vala and granite.
The Flathub repository only lists the Rust version for Gnome systems!